Add JWT authentication for grants-ui service
This commit adds service to service authentication by using a public key
for grants-ui. All endpoints require a token except for the following:

- /health
- /documentation (Swagger)

We are aware the Swagger no longer allows us to try out the endpoint but
we had a few troubles trying to get it to work and wondered how useful it
was actually going to be in practice. Instead we have decided to implement
it as is and if other team members request it then we will try and
implement a solution for swagger.

I tried setting up hapi-auth-jwt2 to dynamically switch between using
jwt auth and no auth depending on if the referer was swagger docs page
but this didn't work. We tried implementing two different strategy to
toggle between depending on if the headers contained the jwt auth token
or a UUID but this didn't work either.